refactor: seperate service
