+++ src/main/java/com/example/CoffeeShop/service/TakeoutService.java
@@ -1,0 +1,11 @@
+package com.example.CoffeeShop.service;
+
+import com.example.CoffeeShop.dto.response.ResponseOrdersDto;
+
+/**
+ * 테이크아웃 서비스 인터페이스
+ */
+public interface TakeoutService {
+    //테이크아웃
+    ResponseOrdersDto takeoutMenu(Long ordersId);
+}
